debuug

+++ price_management/models.py
@@ -23,13 +23,13 @@ class Store(models.Model):
 class StorePrice(models.Model):
     date = models.DateField()
     store = models.ForeignKey(Store, on_delete=models.CASCADE)
-    retail_price = models.FloatField()
-    company_price = models.FloatField(blank=True, null=True)
-    price_1 = models.FloatField(blank=True, null=True)
-    price_2 = models.FloatField(blank=True, null=True)
-    price_3 = models.FloatField(blank=True, null=True)
-    price_4 = models.FloatField(blank=True, null=True)
-    price_5 = models.FloatField(blank=True, null=True)
+    retail_price = models.DecimalField(max_digits=10, decimal_places=3)
+    company_price = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
+    price_1 = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
+    price_2 = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
+    price_3 = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
+    price_4 = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
+    price_5 = models.DecimalField(max_digits=10, decimal_places=3, blank=True, null=True)
 
     def __str__(self):
         return str(self.date)
